test/javax/com/lallen/httpserver/core/response/ResponseHeadTest.java
package javax.com.lallen.httpserver.core.response;
import org.junit.Before;
import org.junit.Test;
import java.io.IOException;

import static org.junit.Assert.assertEquals;

public class ResponseHeadTest {
    public static final String UTF = "UTF-8";
    public static final String ALLOW = "Allow: GET,HEAD,POST,OPTIONS,PUT\r\n";
    public static final String STATUS = "HTTP/1.1 200 OK\r\n";
    public static final String LOCATION = "Location: http://localhost:4040/" + "\r\n";
    public static final String BLANK_LINE = "\r\n";
    public static final String HEAD_LINES = STATUS + ALLOW + LOCATION + BLANK_LINE;
    public ResponseHead responseHead;
    public byte[] header;
    public String head;

    @Before
    public void setUp() throws IOException {
        responseHead = new ResponseHead();
        header = responseHead.renderHead(4040, STATUS);
        head = new String(header, UTF);
    }

    @Test
    public void itGivesTheHeaderInTheCorrectOrder() throws IOException {
        assertEquals(HEAD_LINES, head);
    }
}